[Core] Convert scheduler, playlist, and sequence to C++
This is a simple conversion of the contents of the original
code from schedule.c, playList.c, and sequence.c to C++ classes.
Classes for each source file have been created and the functions
contained in these files have been moved into the classes
along with any global variables defined.  This is an initial
commit with minimal code changes necessary to convert to using
classes.  These changes are primarily to prep for future desired
enhancements to these areas of the code.