[ChannelOutputs] Add LEDscapeMatrix output for Octoscroller
This commit adds backend support for HUB75 LED panels connected
to a BBB via the Octoscroller Cape.  It expects that Debian
2014-05-14 has been installed on the BBB.  This version will be
used for the FPP v2.0 BBB images.  It is available from the
following locations:

http://elinux.org/Beagleboard:BeagleBoneBlack_Debian#2014-05-14
https://s3.amazonaws.com/debian.beagleboard.org/images/bone-debian-7.5-2014-05-14-2gb.img.xz

The data in the sequence should be configured for a horizontal
matrix starting at the top left corner.  Strands per string
should be zero.  A set of two panels in a 2x1 config would be
configured as a 16-string horizontal matrix with 64 pixels per
string and 1 strand per string and starting at the top left
corner.

There is no UI for configuring the LEDscapeMatrix Channel Output
yet, so the /home/fpp/media/channeloutputs config file must be
modified by hand.

The new LEDscapeMatrix config entry should look similar to this:

1,LEDscapeMatrix,1,3072,panels=0:6:N:0:0|0:7:N:32:0

The fields are:

Enabled,OutputType,StartChannel,ChannelCount,OutputOptions

Enabled is 1 or 0 depending on whether the output is enabled or
not.

StartChannel coincides with the start channel in your matrix
model config in your sequencer.

ChannelCount is the total of all channels used on all panels,
which is 1536 per panel for 32x16 RGB panels.

OutputOptions is a set of pipe-delimited panel configs.  In the
above example, there are 2 panels in a horizontal layout.

The OutputOptions panel config fields are:

Output:PanelNumber:Orientation:XOffset:YOffset

Output = Octoscroller output 0-7
PanelNumber = Panel number on Output.  These are numbered 0-7 and
              the highest panel number is closest to the BBB.  The
              panel numbers always start at 7 and count down going
              away from the BBB.  If there are only 2 panels, they
              are still numbered 7 and 6 as in the example above.
Orientation = Direction the top of the panel is facing.
              'N' = Normal/Up
              'L' = Left
              'R' = Right
              'U' = Upside Down
XOffset:YOffset = X,Y offset of top left corner of the panel. The
              top left position in the display is 0,0.  In a
              horizontal config, the panel to the right would be
              at position 32,0 (x = 32, y = 0).  The panel one
              down from the top would be at position 0,16.

Up to 64 panels can be configured, up to 8 on each of the 8
outputs on the Octoscroller board.