[ChannelOutput] Add Direct-attached string support for BBB.
This will be used to drive the F16-B, F4-B, and RGBCape48C capes.

This code uses a new channeloutputs.json file which stores the config
in JSON instead of a flat text file.  The UI for creating this file is
almost done, send me (CaptainMurdoch) a PM if you want the file format
before the UI is ready.