[ChannelOutput] Add UI's for LED Panel matrices and BBB string capes
UI for LED panels supports LEDscapeMatrix/Octoscroller and RGBMatrix
on the Pi.

UI for BBB Cape supports configuring F16-B, F4-B, and RGB123's
RGBCape48C cape.  This may work with the RGBCape24C as well for
the first 24 outputs, but this has not been tested.

This will bypass any existing RGBMatrix or LEDscapeMatrix configurations
in the channeloutputs file.  The UI loads/saves a new
channeloutputs.json file which will be the new config file format
going forward.