[ChannelOutput] Commit Dave's updated pru/FalconWS281x.p source.
This should get F4-B and F16-B working on the BBB.  F4-B has been
tested but strings 2 & 3 are reversed according to the screen print.
Need to check with Dave on whether to fix this in code or screen print
on the PCB.  F16-B has not been tested yet, but will be soon.  F16-B
uses same code as F4-B so it should work.