[ChannelOutput] Add support for USB attached relays
There is no UI for this yet and there may not be for FPP v1.5
because it would require shoehorning in the JSON config file into
the old 'other' channel outputs UI.

This output can control USB attached relays that use the
semi-standard protocol of sending 1 byte per 8 relays where each
bit represents one relay.  bit 0 is the first relay, etc..
There are two different kinds of these that I have seen so far,
the first just takes the raw byte values and is subType 'Bit'.
The second requires an initialization sequence of 0x50 followed
by a pause then 0x51.  One example of these is the relay boards
for sale on icstation.com and also availalbe via eBay.  Model
numbers are ICSE012A for the 4-relay version, ICSE013A for the
2-relay version, and ICSE014A for the 8-relay version.  I have
also seen a 16-port version of these type boards so this code
should support those as well, but they are untested.

Once we redo the UI for FPP v2.0, we can create a setup UI for
this channel output type.  If you want to use one of these relay
boards now, let me know and I can tell you how to hand-configure
the channeloutputs.json config file.