[src] Update for C++ builds with newer gcc
Using newer GCC versions it seems we can no longer use the old C-style
named initialization structures.

In my opinion this is less than ideal because we're not assured the
right fields are set.  This tells me that we should update the
channeloutput and mediaoutput structures to classes and have proper
setter/getter functions, constructors, or something more clean than the
current "curly brace with comment" initialization I've done here.  What
is here now could be prone to error if we update the main definitions of
channeloutput or mediaoutput and don't properly check all instances.