[Makefile] Hack to always update submodules when building
Quick hack to always ensure our submodules are up to date when building.
This will need work to make sure it is functional when there is no
network.  We also might want to set up actual dependencies instead of a
PHONY target so we only run this when there are actually changes to our
submodules.