[Startup] Set hostname in /etc/hosts at startup
Previously we only updated the hosts file when the user updated the
hostname through the UI.  If a user brought over a USB stick onto
a new system, it would boot with the proper hostname, but wouldn't
be in /etc/hosts so sudo would complain.  This commit updates /etc/hosts
when we set the hostname at boot time.