[UI] Change R,G,B to A,B,C on testing page to make it less confusing.
Most users don't have all display items set to RGB order or have all
elements starting on a start channel that is a multiple of 3 minus 2,
so the R,G,B on the testing page was confusing.  For now, until we get
an improved UI, change the R,G,B to A,B,C to make it less confusing.