remove emailgpass value for now
displaying the password is plaintext is a bad idea I know, especially if
someone uploads the a full backup file which contains it to a forum for
dev's to look at. So for now it's nulled out as originally intended by
the ReadSettings function.

I still would like to be able to backup this and be able to restore it
so I don't have to retype my password every time. This is useful when
rebuilding your FPP or have multiple FPP's and you just want to transfer
email settings to them and have it work.

I have an idea for this. Hopefully I can put it down before 1.5 release,
if not it'll it can always be pushed out later.