[UI] Update status page to show when FPP is updating.
Updates can take a while and users sometimes think that FPP is hung
or not starting properly when the automatic update is occurring at
boot time.  This change will set the status to "FPP is currently
updating" when git_pull.sh is running instead of showing
"FPPD is stopped" for the duration.