Added option for user to be able to export sensitive data in backup
By default passwords (currently only email acc.) are not included in the
backup for obvious reasons.

There is an option to export this data via the backup page, with a
warning about passwords being stored in plain-text. The backup filename
also changes to include the wording "unprotected".

Slight rewording of the backup and restore instructions.