[Logging] Add locking around logging
This is a quick fix to add locking around the non-threadsafe
functions used in the logging function.  This can be refined
later to minimize the locked area, but at our normal log level
the impact should be minimal since the locked area does not
include the code to check whether we should log or not.