[ChannelOutput] Fix RPIWS281x channel output on the Pi.
There is still no UI for this, but it can be tested by adding something
like the following to your /home/fpp/media/channeloutputs config file:

1,RPIWS281X,1,60,string1GPIO=18;string1Pixels=20

The fields are:

1 - enabled
RPIWS281X - channel output type
1 - start channel
60 - channel count
string1GPIO=18;string1Pixels=20 - channel output options.

The code should support two independent strings by specifying a
second GPIO pin and string pixel count such as this:

string1GPIO=18;string1Pixels=20;string2GPIO=19;string2Pixels=20

The channels for the second string must immediately follow the channels for
the first string.

This code does not work on the Pi v2 B yet, there is a known issue in
the rpi_ws281x library we are using and it will need to be patched or
we can possibly switch to a patched fork.