[MediaOutput] Convert TagLib metadata parsing to use C++ classes
There may be a crash issue related to memory corruption somewhere in
the C metadata parsing.  This has been showing up on the quad-core
Pi v2 B.  The TagLib C code was used before we started compiling with
g++, but now that we are using g++ we can convert to TagLib C++ classes.
Hopefully this will fix the crash issue as well.