[ChannelOutput] Add X11Matrix test channel output.
This channel output will open a X11 Window and use it as a RGB Matrix
for testing other parts of FPP code.  For testing a low-density matrix,
a scaling factor can be applied so for example a 32x16 matrix can be
drawn as 320x160 using a scale of 10 to create 10x10 square "pixels".

The config is in channeloutputs.json and looks like this:

        {
            "type": "X11Matrix",
            "enabled": 1,
            "startChannel": 1,
            "channelCount": 6144,
            "width": 64,
            "height": 32,
            "scale": 10
        },