[Build] Add ability to include an optional local Makefile.local file
This can be used to override or append to values set in the Makefile
included in git.

One example of this is enabling the X11Matrix Channel Output for
debugging purposes using the following lines in Makefile.local:

    LIBS_fppd += -lX11
    CFLAGS += -DUSE_X11Matrix
    OBJECTS_fppd += channeloutput/X11Matrix.o

Other uses might be to compile some Pi-specific code on a non-Pi for
compile testing even if the code could not be used on a non-Pi.