[MultiSync] Sync blanking data between Master and Remotes.
Previously, the Remote would send blanking data whenever a sequence
ended.  This has been removed and now the Master sends a message to
the remote to tell it to send blanking data if it needs to.  This
fixes the problem of the display going dark on remotes in between
sequences.
(cherry picked from commit a5f075a394faeda243ea794f9091a91300d1c3f8)

Conflicts:
	src/Sequence.cpp