[scripts] Update image to 17 to fix dns issues
On the Pi the openresolv package has if-up/if-down scripts that will end
up overwriting the /etc/resolv.conf script when the interfaces change.
We have seen cases where these are run even when the interface appears
to maintain connectivity.

The fix:

1. Add a "Generated by fpp" comment to the /etc/resolv.conf that
signifies that we are using manual DNS generated by FPP.

2. Update the if-up/if-down scripts from openresolv to check for the
"Generated by fpp" line and don't run resolvconf to prevent it from
overwriting our config file.