[ChannelOutput] Upgrade RGBMatrix library for Pi LED panel support.
This new code uses the hardware PWM on the Pi for better performnace
but unfortunately that means it conflicts with the onboard audio
on the Pi.

I want to look at allowing the code to be compiled without hardware
PWM support, but for now, it is necessary to disable onboard audio
when the RGBMatrix Channel Output code is turned on.