[Playlist] Initial commit of partially-working playlist conversion
This is the first commit of the work-in-progress playlist conversion
to C++.  THIS IS NOT FULLY FUNCTIONAL YET.

The playlist UI has been modified to save the playlist in both the
original CSV format and the new JSON format to make testing and
development easier.

The following playlist entry types are working or partially working:
- effect (new type, not in UI)
- event
- media
- pause
- pixelOverlay (new type, not in UI)
- volume (new type, not in UI)

The following playlist entry types are not working at all yet:
- branch (new type, not in UI)
- media + sequence
- playlist (new type, not in UI)
- plugin
- sequence

There is still a lot of work to be done on this conversion, but the
code is being committed to make future development easier.