[scripts] Fix startup script that wouldn't configure and start apache2
When moving changes from a Pi install to my development machine, somehow
tabs for a specific grep line were converted to spaces causing the
apache detection in our bootup script to fail.  This would attempt to
configure nginx instead of apache, breaking older installs after a
reboot.

This should hopefully fix that issue and continue to support the old
apache2 installs as intended.