[Settings] Output better name for a USB-FM Transmitter
I purhcased a couple of these modules a while back based on a thread on
diychristmas.org and finally got around to plugging one in.  It "just
works" with the name "CD002" in our drop-down since that's what it
identifies itself to alsa as.

http://www.aliexpress.com/item/Brand-New-Top-Quality-DSP-PLL-87-108MHz-Digital-Wireless-Microphone-Stereo-FM-Transmitter-Module/32382137417.html

The volume control doens't work, I'll have to look at what the name of
the mixer is to change our volume setting commands when this device is
selected.
(cherry picked from commit 97231ad358e26a608c25dd5739f2716c7f8d993a)