[network] FPP Access Point
To use, in your interfaces.wlanX config file set PROTO=master to enable.
For the ADDRESS=/NETMASK= fields set the IP subnet to use. The script will
set the interface IP to the first address of the specified subnet regardless
of what's the ADDRESS field sya . So if you set ADDRESS=192.168.10.5 and
NETMASK=255.255.255.0, the script will calculate the NETWORK as 192.168.10.0,
give the FPP the address of 192.168.10.1 and if DNSMASQ=yes, setup a DHCP range
of 192.168.10.2 to 192.168.10.254. If you specifiy NAT_IFACE= with a interface name,
the script will enable outbound NAT from this interface to the specified interface.
NAT_IFACE is optional, if not specified or blank it will not configure any NAT. DNSMASQ
is optional, but on by default. To disable DNSMASQ, set DNSMASQ=no.

Additional options are CHANNEL= and COUNTRY=, which if not specified default
to CHANNEL=6 and COUNTRY=US. CHANNEL is the 2.4GHZ WiFi channel and COUNTRY is
ISO Country Code to setup up the correct regulatary domain for the WiFi.

SSID and PSK work the same as before, if PSK is not specified the access point
will operate in the clear. The GATEWAY value is ignored when PROTO=master.

These additional packages are required:

wireless-regdb iw crda ipcalc haveged hostapd dnsmasq dnsmasq-base libnetfilter-conntrack3

Please make sure to disable hostapd and dnsmasq from starting automatically
at boot time.

The PROTO=master, DNSMASQ=, NAT_IFACE=, CHANNEL=, COUNTRY= are not enabled in the
GUI so these options must be set by manually editing the interface file.
It also may not be obvious but "Enable Routing" must also be checked in the GUI.