[Playlist] First commit for new Playlist in v1.x tree
The old playlist code is still in here, but commented out or unused.
It is slowly being ripped out after regression testing is performed.

This commit also includes a new MQTT client to allow FPP to publish
and subscribe to MQTT topics.  Currently minimal publish messages
are sent, but more will be added over time.  There is a new Playlist
entry which allows publishing a MQTT message directly from a playlist.

In the playlist code, the following items are believed working, but
still need a lot of regression testing:

- Media
- Sequence
- Media+Sequence
- Event
- Script
- MQTT
- Branch

Known Issues:
- The 'Plugin' Playlist entry type is not working at all, it is not
  hooked up currently.