[Playlist] Add 'Dynamic' playlist entry type.
The Dynamic type reads and parses a JSON format string that contains
a single playlist entry.  This entry is then run as if it were part
of the original playlist .json file.  The next time through the playlist
loop, the data source is re-read and re-parsed, allowing for dynamic
playlists driven from outside sources.