[ChannelOutput] Add VirtualStrings support for RPIWS281X Channel Output
RPIWS281X output moves to a new tab on the Channel Output config
screen and will have to be reconfigured by the user.

Brightness/gamma is not currently hooked up until Dan's BBB-updates
branch is merged back to master-v1.x and then to master.

Needs more sanity checks in the UI as well such as max # of pixels per
physical port.