[Core] Share /home/fpp/media instead of /home/fpp
Share /home/fpp/media so that Samba reports the correct amount of
disk space and free space for the storage media instead of for
the main /home/fpp directory.  Restart Samba in the fpp startup
script after the storage media is mounted to ensure that Samba is
sharing the correct directory.

Fixes #301.