[ChannelOutput] Convert Channel Remap config to JSON
Also adds in the ability to set a remap as active/inactive rather
than having to delete a remap to disable it.  Adds a description
field to allow identifying what a remap is for, and also adds a
'loop' count so a single set of source channels can be remapped
multiple times at a destination.  The loop functionality is
similar to grouping functionality on a pixel controller, with a
loop count greater than 1, a single set of 3 RGB channels could be
remapped onto a set of multiple pixels at the destination channel.

Fixes #183.