[MediaOutput] Compile in SDLOutput support on all platforms.
For now, the SDLOutput class must be forced On via a hidden setting
until the known issues are worked out.  We still need to get the
SDLOutput to close when finished or stopped so that it doesn't
lock the audio device.

To Enable SDLOutput usage, create a setting called "ForceSDL" with a
value of 1 in the FPP settings file.  Eventually this setting will
be removed and the LegacyMediaOutputs setting will be uncommented to
allow users to fall back to the legacy mpg123/ogg123 outputs if
necessary.

NOTE: This commit requires libsdl2-dev which is installed via the
      master FPP_Install.sh script or the upgrade/28/upgrade.sh script.

References #310.