[Boot] Remove volume adjustment workaround and fix issue with NFS storage
We no longer need the workaround for the volume adjustment at boot.  The
firmware issue was fixed a long time ago as referenced in
https://github.com/raspberrypi/linux/issues/570

When using a NFS share for /home/fpp/media, the storage source location
was not getting parsed properly due to the length of the field, so change
the parsing of the findmnt output a little to handle this case.