[Scripts] Fix unnecessary warning when shutting down fppd
fppd_stop tells fppd to quit using "fpp -q".  Previously, fpp would
exit and it was a race to see if fppd stopped first or fppd_stop
checked to see if fppd was stopped, so we added a sleep in the script.

This commit updates the "-q" command parsing to sleep for one second.
This gives fppd time to shutdown and actually causes the fpp socket
connection to be closed by fppd terminating.  This means we can
shorten up the sleep time.  The change also includes added code to
check to see if fppd is actually still running instead of always
calling killall which would generate an error if fppd was not found.