Update Logs Zip functionality
Update Logs -> Zip to include new config files introduced in FPP v2.

Made a small tweak to the file reading so 'schedule',
'channelmemorymaps', 'channeloutputs', 'channelremap', 'universes' are
read in correctly. ScrubFile assumes a INI file if there is no .json
extension and it breaks for schedule, the other files don't need to be
scrubbed anyway