Multisync Pagehang / Stops responding fix
Issue is to do with PHP sessions and session locking. nginx runs scripts
super quick and we end up with read locks on the php session. This is
worse when you have alot of FPP devices that need polling, we end up
with 6+ get getFPPStatus requests all firing off at once.

Fix is to close the session before we start any FPPStatus work (we don't
need it anyway) and at the beginning of resturnJSON.

Added in minor handling of errors with file_get_contents
Added reason for a failed read when the remote device has a WebGUI
password via a 'unknown' status

Will backport this into v1.x as it was happening there, rarely but it
did happen - also session related