Fix playlist delete function
Added a check to see if file exists first, if it does it's likely the
user is deleting a older v1 playlist (CSV without a extension).

If the file doesn't exist, append .json onto the filename we're supplied
and try to delete. All v2 playlists are json files with the approriate
extension.
Also added ability to return failure XML status for playlist deletes.

Added a popup dialog when playlist deletes fail