Fix Issue adding events to Playlists
When adding events to playlists, no event name is shown instead 'undefined' is displayed. Events will save into the playlist but cause futher issues.

Event majorID and minorID's are being saved as strings in the playlist json, also a leading 0 is present for id's < 10. FPPD seems to crash if these are left as strings.

Added logic to keep these as integers and keep the leading 0 off low numbers.
Some additional fixes to handle event ID's

Fixes #397