Fix issue saving playlists with media of filenames containing '%'
Fix issue saving playlists with media of filenames containing '%'
Bit a edge case, users may run into this if they have weird filenames on MP3's also.

Swapped fprintf with fwrite to write out the playlist JSON string as the '%' sign used in a string with fprintf denotes the start of a format specifier. And it tries format the string and insert something into that spot.

Added a little check for empty JSON content

Fixes #419