Enhancement - File Manager display filesizes & media durations
Small enhancement to the File Manager to now show the filesizes of sequences, effects etc. and the duration of music and video files.

Implements a cache for the media file durations since retrieving this from the files consumes a very small amount of time per file about 100ms/file, but in bulk it adds up to a few seconds.
Cache persists for a very long time, 30 days, filesizes are used to help refresh any changes to individual files during that time.