[Scheduler] Fix bug causing manual playlists to stop prematurely
This reverts part of 96b26e0af.  The Scheduler really doesn't know
about manually run playlists, so we have no current way to try to
stop a playlist other than at the scheduled end time.

This bug was causing manual repeated playlists to stop gracefully
unless there was a playlist scheduled to be run after the current
time before the end of the week.