Detect if the time has jumped into the future and reschedule.
Sometimes when the Pi comes up, the date may be well in the past
and the Scheduler may not detect the next scheduled run time
properly if the date is then corrected instead of passing the
midnight boundary when a reschedule is forced.  This patch will
reschedule if the time jumps more than 5 seconds.