Add logic to the Playlist so it knows whether it was scheduled or not.
This fixes the new bug of manually started playlists immediately going
into 'stopping gracefully' mode.  The scheduler now knows when a
playlist was manually started and doesn't stop it due to bad data.

This also is part of the groundwork necessary to fix another couple
issues related to manually started playlists.  Adding those to my ToDo.