Update RPIWS281x Channel Output to allow testing new library code
The new version of the rpi_ws281x library supports 2 additional string
outputs using PCM and SPI.  This commit makes it easier for us to test
FPP with this new code using a manually updated co-pixelStrings.json
file.

So far, the SPI mode has been tested, but with a very observable
flicker while also using the original PWM outputs.  This could be
something else in FPP messing with the SPI device so we may need
to update FPP to disable other code to use the additional outputs.
Or, it could be that you can't use both PWM and SPI at the same
time which means we're limited to a max of two strings.  I haven't
tested PCM output because initial testing was done on a Pi v1 B
which does not have the required PCM pin.

To test this code, setup the Pi Pixel Strings Channel Output with
the second string set to 0 pixels.  Edit
/home/fpp/media/config/co-pixelStrings.json and add a "gpio" key
in the first output below the portNumber key.  Set the gpio value to
10 to test the SPI output using pin 19 on the Pi.  Set the gpio value
to 21 to test the PCM output using pin 40 on the Pi.