Add a HTTP Virtual Display Channel Output
This channel output is used for displaying a Virtual Display in a
HTML5 Canvas in a web browser.

You must have a virtualdisplaymap CSV file installed and have
uploaded your xLights preview image as a file called
virtualdisplaybackground.jpg to use this plugin.  There currently
is no UI to set this up since we need to get the virtualdisplaymap
patch applied to xLights to allow generation of the CSV file containing
the pixel locations.