Allow building binaries in a separate subdir for development.
The main use for this patch is to allow compiling multiple
architectures from a single source.  For instance, I can mount
my fpp dev directory from my dev machine onto a Pi as /home/fpp/fpp
and compile in /home/fpp/fpp/src on the Pi for ARM while compiling
for x86_64 in the 'build' directory on my dev machine using the same
source files.  This makes it easier to regression test changes
on the Pi/BBB while getting fast compiles on a faster dev system.
All .o files and executables would be created under the build
directory.

Usage, assuming you are in the src directory on a machine, is:

cd ../
mkdir build
cd build
make -f ../src/Makefile

After that, the fpp, fppmm, fppd, etc. binaries should exist in
the build directory.  The build directory is in the .gitignore
file, so these files will be ignored by git.