Initial version of PlaylistEntryImage for displaying images from a playlist
This new playlist entry type will allow displaying an image on the
Pi/BBB HDMI/composite output.  This could be used for a TuneTo sign or
with a projector.

The UI for adding an image entry to a playlist is currently disabled while
I test further and implement the image cache cleanup code.

Also, use CXXFLAGS when compiling .c files.  We are already compiling .c
files with g++, so we may as well pass in the same flags as we use when
compiling .cpp files.