Add support for sunset/sunrise when scheduling playlists.
Users can now select 'SunRise' or 'SunSet' in the schedule
entry times to start/end a playlist at sunrise or sunset.

Uses the SunSet class from https://github.com/buelowp/sunset
which is licensed under the MIT license.  License is noted in
the SunSet.cpp and SunSet.h source files.

SunRise is stored as 25:00:00 in the schedule config file and
SunSet is stored as 26:00:00.  Once we move this CSV file to
JSON, we can have a better storage method.