Prevent some special characters being used in playlist name
Attempt to somewhat sanitize Playlist names that have special characters in them, mostly commas. This is to prevent issues in the scheduler since it's CSV based at current.
This is purly a backend workaround/fix since we want to fix the name before it's used in the php session and later when it's referenced in the php session.
Once saved the playlist is saved with the commas replaed with hyphens ( - )

Closes #476

(cherry picked from commit 618bc76)