A few tweaks to allow installation on stock Debian on non-Pi/non-BBB.
Also clean up /root/.cpanm to save SD image space, and fix the copying
of the FPP_Install.log and .sh file to /home/fpp.

fppoled still doesn't build since gpiod isn't included in Stretch, but
it looks like it may be in Buster so it should build once we start
using that OS version.