Allow playlist nesting 5 levels deep and m_scheduled flag bugfix
Honor m_scheduled flag when inserting a playlist so the scheduler
knows whether the playlist was scheduled or not after we swap
playlists.  I haven't heard of any bugs reports related to this,
but it appears it could have caused a playlist not to stop on time
if the inserted playlist was running during the scheduled stop time.