Fix some flicker when scrolling text in Pixel Overlay model
When scrolling text across the Pixel Overlay model, the channel
output code would sometimes grab the frame in the middle of the
clear() call.  This was causing occasional flicker.  Now, we
double buffer and only need to clear our buffer before drawing on
it again and then copying the whole buffer into place.  This still
doesn't fix 'tearing' that may occur because the scroll is not gen
locked to the channel output code, but it is better than before.