Fix scheduler bug regarding overlapping playlists.
When the scheduler normally runs, it will play the first item it
finds in the list if one or more items overlap.  This patch fixes
an issue which allowed fppd to try to start both playlists if it
started up during the overlapping time period.  The second would
stop the first and the second one would be run.  This makes this
situation different than the normal scheduler run, so return
as soon as we find the first item that matches.