Add ability to treat a Pixel Overlay Model as a RGB light in Home Assistant
/falcon/player/{HostName}/light/{ModelName}/cmd
/falcon/player/{HostName}/light/{ModelName}/state

These take and send the normal JSON schema light messages for the light.mqtt
integration in Home Assistant.

https://www.home-assistant.io/integrations/light.mqtt/#json-schema

{"state": "ON", "color": {"r": 0, "g": 0, "b": 255}, "brightness": 255 }

brightness and color fields are optional, if they are not specified, then the
last cached values will be used.  If there is no cached value for that model,
then 255 will be used for each value.