Remove auto-update functionality
Auto Update has been disabled for a couple years now and we were not
recommending that users enable the setting, so the setting is being
removed.  If users still desire to auto update, it can be easily done
by adding a few lines of code to the User Callback hook script in
the script repository.

Fixes #664