Add a test mode for GPIO Input/Output code
This patch allows testing the GPIO Input, Output, and UI code on
non-Pi/non-BBB/non-Docker systems.

When compiled on non-Pi/non-BBB/non-Docker systems, a
TmpFilePinCapabilities class will be used which similates GPIO Inputs
and Outputs using files in /tmp which is normally a tmpfs filesystem and
very quick.  When used as an Output, a file will be created and removed
when one of these virtual GPIOs is turned on and off.  When used as an
input, the GPIO will be considered On if the file exists and considered
Off when the file is not present.  Twenty virtual GPIO inputs are
simulated using files named /tmp/GPIO-TF-1 up to /tmp/GPIO-TF-20