Create/expand set of JSON load/save helpers
Several of the jsoncpp functions we use to read and write JSON
are being deprecated.  We also have lots of duplicated code to
call these functions.  This patch adds a set of helper functions
to allow us to migrate off the deprecated functions and simplify
some of the JSON read/write code throughout FPP.

This is the first part of a two part patch.  The first adds the
basic helpers and the second will migrate the remaining portion
of FPP to use the new helpers.  The second patch will be committed
once it undergoes a little more regression testing.