Fix Playlist editor for strings containing HTML encoded characters
Specifically, &amp; was being written to the playlist .json file
but there are probably other characters which shouldn't have been
stored as their HTML encoded version.  This manifested itself as
an error in the player indicating that the file was not found.

To fix an existing playlist, load the playlist in the editor and
save it again.

Fixes #872.