Don't reprocess cached channel data after sequence ends.
Issue discovered on system running brightness plugin with
a repeating overnight sequence.

This was manifesting itself as a blanking data issue when in
fact the issue was the brightness plugin being called multiple
times on the same channel data after the sequence finished and
before it was opened again causing the brightness of the cached
data to go to near zero and the lights to seemingly flicker off
briefly until the sequence was reopened and fresh data was read.

I don't think there are any open bug reports about this, but it
may solve issues some users have seen related to blanking
between sequences.