Updated JSON Backup to gather extra JSON config files & network interfaces
Updated the JSON settings backup system to gather and pull in any extra JSON config files found in the /config folder, some filtering is done so that we don't pull in any config files which are already mapped to know areas (eg say Command Presets).
This change should enable the system to be a little more capable and dynamic in handling config files we may not know about.
Extra checks are also done to ensure the file contents is actually JSON just to be sure.

Updated backing up of network interfaces, a similar approach to pulling in extra JSON files is taken so that all configured interfaces are backed up so users with extra USB ethernet or WiFi interfaces will have configs backed up correctly

Users who are relying on the JSON backup system should be prompted to generate/download a new backup after this change goes out just so they have a good backup source (without any missing settings)