more tweaks to work around the VLC freezing issues
many changes to help, but not completely eliminate, the video playback freezes.  Most of the changes are additional and enhanced debugging.  Other changes include re-ordering some of the logic to deal with an issue when flipping between fast-playback and slow-playback.  There was a race condition where we initially played faster to catch up and we would slow down as we got close, but never got down to normal speed so it kept creeping slowly further ahead until it tried to 'jump' which causes a video freeze in VLC.  Similar condition happened when going slower than normal rate, but was much less common.   Also increased maximum speedup to double-speed and minimum slowdown to half-speed and upped the 'wait-to-seek' threshold to 10 seconds (abnormally long).  We should never be that far off unless fppd restarts.  The 10 seconds is a hack but it allows FPP more time at 2x speed to get back in sync before trying to do a 'seek' which will trigger a VLC video freeze.