Schedule out 5 weeks to provide better preview of future actions
The new time_t based scheduler allows scheduling out into the future
an arbitrary number of days.  Up until now, that has been limited to
7 days counting the current date.  This patch changes the scheduler
to schedule out 5 weeks in advance to allow the schedule preview
and "Next Playlist" field on the main status screen to show items
a month into the future.