Fix bug when schedule entry end time is set to start time
Previously, if a schedule entry was set to repeat and the end
time was set to the same as the start time, the entry would not
get scheduled to run at all, even if the stop type was set to
graceful loop.  The desired action in that case would be to run
the item once and let it end when it finishes according to the
stop type.