Remember playback position when restarting during scheduled playlist
When fppd performs a quick restart during a scheduled playlist, exec
the new fppd instance with command line args to allow the new fppd
instance to resume the playlist at the same position.  This will
restart the current item, so restarts are best performed immediately
after a new sequence or media file is started.