Use scheduler to restart playlists that should be playing
If the user starts a playlist manually that is supposed to be
running via the scheduler, use the scheduler to start the playlist
so that it stops at the scheduled end time.  This was a regression
in the change to using the REST API instead of the old command
socket code.

Fixes #1196