Fix a race condition in playlist startup.
The Player was calling the Playlist::Play() and ::Start()
routines separately allowing a race condition to occur if the
main loop called Player::Process() -> Playlist::Process() in
between the Play() and Start() calls.  This would cause an
error in the logs about an invalid section position and the
playlist would be immediately stopped before the Start() call.

References #1159