Fix second-midnight not working in scheduler UI in 12-hour mode
For years we have told people to use 00:00:00 to 24:00:00 to
schedule a playlist that will start whenever FPP starts.  When the
new UI is in 12-hour time mode, this 24:00:00 time was showing
as '12:00:00 AM' and getting saved as 00:00:00 which broke this
functionality.  This commit upgrades the jquery timepicker
JavaScript file, switches to the non-minified version, and also
includes a FPP patch to show this 'second midnight' as
'12:00:00 Mid' along with FPP changes to properly save this
time as 24:00:00 internally to restore the original functionality
for users using the 12-hour UI mode.

Fixes #1207